Support temporary clients
Temporary clients can be generated by reading the instance metadata on a
given node. However, such clients have the caveat that their requests
require a special token under the header name X-Amz-Security-Token.

This patch makes it so temporary clients can be created with an explicit
security token passed in, and that the signature of headers includes
this token as a header when required.