Add support for a temporary client
AWS provides a local DynamoDB for development.
Requests to this dynamodb need to be made to
localhost, over unencrypted HTTP. It can be
reasoned that if AWS is to release other services
for local development it will follow a similar
pattern.

This commit provides a local client that connects
to localhost.

PR for the code generator will follow soon.